backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "beanie>=2.0.0",
    "fastapi>=0.120.4",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.3",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "uvicorn>=0.38.0",
]
